refactor: searchページのviewとlogicの分離
- searchページのlogicとviewをファイル分けした
