+++ src/app/user/page.tsx
@@ -1,5 +1,4 @@
 import PostPageLogic from "./logic/Logic";
-// 本来はロジックのコンポーネントが来る
 
 export default function Page() {
   return <PostPageLogic />;
